Remove the push Registration Screen
