+++ app/src/main/res/layout/fragment_push.xml
@@ -8,46 +8,9 @@
         android:id="@+id/messages"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        app:layout_constraintBottom_toTopOf="@id/refreshToken"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/refreshToken"
-        android:layout_width="0dp"
-        android:layout_height="60dp"
-        android:text="Refresh Token"
-        android:enabled="false"
-        android:background="@color/colorPrimary"
-        android:textColor="@color/colorPrimaryDark"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintBottom_toTopOf="@id/register"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/register"
-        android:layout_width="0dp"
-        android:layout_height="60dp"
-        android:text="Register"
-        android:background="@color/orange"
-        android:textColor="@color/white"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/unregister"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/unregister"
-        android:layout_width="0dp"
-        android:layout_height="60dp"
-        android:enabled="false"
-        android:text="Unregister"
-        android:background="@color/orange"
-        android:textColor="@color/colorPrimaryDark"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/register" />
-
 
 </android.support.constraint.ConstraintLayout>
